requirements.txt
# pip install -r requirements.txt

PyJWT
flask
flask-sqlalchemy
psycopg2-binary
python-dotenv
